+++ helloweb Maven Webapp/src/main/java/cn/zhouyafeng/blog/dao/CommentDetailEntityMapper.java
@@ -1,6 +1,6 @@
 package cn.zhouyafeng.blog.dao;
 
 public interface CommentDetailEntityMapper {
-	public int getNextCommentId();
+	public long getNextCommentId();
 
 }
